+++ coloraide/spaces/oklch/__init__.py
@@ -33,6 +33,7 @@ from ... import algebra as alg
 from ...types import Vector
 
 ACHROMATIC_THRESHOLD = 0.000002
+ACHROMATIC_HUE = 90.00000025580869
 
 
 def oklab_to_oklch(oklab: Vector) -> Vector:
@@ -55,6 +56,15 @@ def oklch_to_oklab(oklch: Vector) -> Vector:
     """OkLCh to Oklab."""
 
     l, c, h = oklch
+
+    # For better round tripping of achromatic colors,
+    # use the achromatic hue that occurs in forward transform.
+    # We use the one from white translation. It may or may not vary slightly
+    # depending on the grayscale color, but only slightly,
+    # so this is close enough.
+    if c < ACHROMATIC_THRESHOLD:
+        h = ACHROMATIC_HUE
+
     if alg.is_nan(h):  # pragma: no cover
         return [l, 0.0, 0.0]